Move Transaction Repository into interface package and updated the Transaction package to have a one-to-many relationship within the database
