+++ backend/src/main/java/com/bank/backend/interfaces/TransactionRepository.java
@@ -1,5 +1,6 @@
-package com.bank.backend.transaction;
+package com.bank.backend.interfaces;
 
+import com.bank.backend.transaction.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
