+++ backend/src/test/java/com/bank/backend/transactionTest/TransactionRepositoryTest.java
@@ -1,17 +1,9 @@
 package com.bank.backend.transactionTest;
 
-import com.bank.backend.bankaccount.BankAccount;
-import com.bank.backend.bankaccount.BankAccountRepository;
-import com.bank.backend.transaction.Transaction;
-import com.bank.backend.transaction.TransactionRepository;
-import com.bank.backend.transaction.TransactionType;
+import com.bank.backend.interfaces.TransactionRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class TransactionRepositoryTest {
